x/specy/keeper/query_current_executor_status.go
package keeper

import (
	"context"

	sdk "github.com/cosmos/cosmos-sdk/types"
	"github.com/specy-network/specy/x/specy/types"
	"google.golang.org/grpc/codes"
	"google.golang.org/grpc/status"
)

func (k Keeper) CurrentExecutorStatus(goCtx context.Context, req *types.QueryGetCurrentExecutorStatusRequest) (*types.QueryGetCurrentExecutorStatusResponse, error) {
	if req == nil {
		return nil, status.Error(codes.InvalidArgument, "invalid request")
	}
	ctx := sdk.UnwrapSDKContext(goCtx)

	val, found := k.GetCurrentExecutorStatus(ctx)
	if !found {
		return nil, status.Error(codes.NotFound, "not found")
	}

	return &types.QueryGetCurrentExecutorStatusResponse{CurrentExecutorStatus: val}, nil
}
